landing_backend/requirements.txt
annotated-types==0.7.0
azure-core==1.30.1
azure-cosmos==4.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.0
urllib3==2.2.1
